fOOD_HUB_ORDER
The food aggregator company has an internet site where they save the information of the various orders placed by registered clients. In order to improve the dining experience for its patrons, they wish to examine the data to obtain a clear understanding of the demand from various eateries. Let's say this organization hires you as a data scientist, and the team has highlighted some of the important questions that need to be addressed. To uncover the answers to these queries that will enable the business to grow, conduct data analysis.